requirements.txt
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.34.2
httpx==0.24.0

# Google Sheets integration
google-api-python-client==2.95.0
google-auth==2.21.0